+++ lib/sidekiq_unique_jobs.rb
@@ -30,8 +30,7 @@ require "sidekiq_unique_jobs/batch_delete"
 require "sidekiq_unique_jobs/delete_orphans"
 require "sidekiq_unique_jobs/cli"
 require "sidekiq_unique_jobs/core_ext"
-require "sidekiq_unique_jobs/timeout"
-require "sidekiq_unique_jobs/timeout/calculator"
+require "sidekiq_unique_jobs/time_calculator"
 require "sidekiq_unique_jobs/unique_args"
 require "sidekiq_unique_jobs/unlockable"
 require "sidekiq_unique_jobs/key"
