+++ lib/sidekiq_unique_jobs/script/caller.rb
@@ -49,6 +49,8 @@ module SidekiqUniqueJobs
         end
       end
 
+      # Only used to reduce a little bit of duplication
+      # @see call_script
       def do_call(file_name, conn, keys, argv)
         argv = argv.dup.concat([now_f, debug_lua, max_history, file_name])
         Script.call(file_name, conn, keys: keys, argv: argv)
