+++ lib/sidekiq_unique_jobs/testing.rb
@@ -6,7 +6,17 @@
 require "sidekiq"
 require "sidekiq/testing"
 
+#
+# See Sidekiq gem for more details
+#
 module Sidekiq
+  #
+  # Temporarily turn Sidekiq's options into something different
+  #
+  # @note this method will restore the original options after yielding
+  #
+  # @param [Hash<Symbol, Object>] tmp_config the temporary config to use
+  #
   def self.use_options(tmp_config = {})
     old_config = default_worker_options
     default_worker_options.clear
@@ -18,8 +28,21 @@ module Sidekiq
     self.default_worker_options = old_config
   end
 
+  #
+  # See Sidekiq::Worker in Sidekiq gem for more details
+  #
   module Worker
+    #
+    # Adds class methods to Sidekiq::Worker
+    #
     module ClassMethods
+      #
+      # Temporarily turn a workers sidekiq_options into something different
+      #
+      # @note this method will restore the original configuration after yielding
+      #
+      # @param [Hash<Symbol, Object>] tmp_config the temporary config to use
+      #
       def use_options(tmp_config = {})
         old_config = get_sidekiq_options
         sidekiq_options(tmp_config)
@@ -30,6 +53,9 @@ module Sidekiq
         sidekiq_options(old_config)
       end
 
+      #
+      # Clears the jobs for this worker and removes all locks
+      #
       def clear
         jobs.each do |job|
           SidekiqUniqueJobs::Unlockable.unlock(job)
@@ -40,25 +66,19 @@ module Sidekiq
       end
     end
 
+    #
+    # Prepends deletion of locks to clear_all
+    #
     module Overrides
-      def self.included(base)
-        base.extend Testing
-        base.class_eval do
-          class << self
-            alias_method :clear_all_orig, :clear_all
-            alias_method :clear_all, :clear_all_ext
-          end
-        end
-      end
-
-      module Testing
-        def clear_all_ext
-          clear_all_orig
-          SidekiqUniqueJobs::Digests.new.del(pattern: "*", count: 1_000)
-        end
+      #
+      # Clears all jobs for this worker and removes all locks
+      #
+      def clear_all
+        super
+        SidekiqUniqueJobs::Digests.new.del(pattern: "*", count: 1_000)
       end
     end
 
-    include Overrides
+    prepend Overrides
   end
 end
