--- lib/sidekiq_unique_jobs/timeout.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module SidekiqUniqueJobs
-  module Timeout
-  end
-end
