+++ lib/sidekiq_unique_jobs/web/helpers.rb
@@ -2,21 +2,45 @@
 
 module SidekiqUniqueJobs
   module Web
+    #
+    # Provides view helpers for the Sidekiq::Web extension
+    #
+    # @author Mikael Henriksson <mikael@zoolutions.se>
+    #
     module Helpers
-      VIEW_PATH = File.expand_path("../web/views", __dir__)
+      VIEW_PATH    = File.expand_path("../web/views", __dir__)
+      SAFE_CPARAMS = %w[cursor prev_cursor].freeze
 
       module_function
 
+      #
+      # Opens a template file contained within this gem
+      #
+      # @param [Symbol] name the name of the template
+      #
+      # @return [String] the file contents of the template
+      #
       def unique_template(name)
         File.open(File.join(VIEW_PATH, "#{name}.erb")).read
       end
 
+      #
+      # The collection of digests
+      #
+      #
+      # @return [SidekiqUniqueJobs::Digests] the sorted set with digests
+      #
       def digests
         @digests ||= SidekiqUniqueJobs::Digests.new
       end
 
-      SAFE_CPARAMS = %w[cursor prev_cursor].freeze
-
+      #
+      # Creates url safe parameters
+      #
+      # @param [Hash] options the key/value to parameterize
+      #
+      # @return [String] a url safe parameter string
+      #
       def cparams(options)
         # stringify
         options.keys.each do |key|
@@ -30,6 +54,13 @@ module SidekiqUniqueJobs
         end.compact.join("&")
       end
 
+      #
+      # Redirect to with falback
+      #
+      # @param [String] subpath the path to redirect to
+      #
+      # @return a redirect to the new subpath
+      #
       def redirect_to(subpath)
         if respond_to?(:to)
           # Sinatra-based web UI
@@ -40,17 +71,38 @@ module SidekiqUniqueJobs
         end
       end
 
+      #
+      # Gets a relative time as html
+      #
+      # @param [Time] time an instance of Time
+      #
+      # @return [String] a html safe string with relative time information
+      #
       def relative_time(time)
         stamp = time.getutc.iso8601
         %(<time class="ltr" dir="ltr" title="#{stamp}" datetime="#{stamp}">#{time}</time>)
       end
 
+      #
+      # Gets a relative time as html without crashing
+      #
+      # @param [Float, Integer, String, Time] time a representation of a timestamp
+      #
+      # @return [String] a html safe string with relative time information
+      #
       def safe_relative_time(time)
         time = parse_time(time)
 
         relative_time(time)
       end
 
+      #
+      # Constructs a time from a number of different types
+      #
+      # @param [Float, Integer, String, Time] time a representation of a timestamp
+      #
+      # @return [Time]
+      #
       def parse_time(time)
         case time
         when Time
